refactor(ui): extract 5 leaf settings components into settings/components/
- SettingsSectionIntro.tsx (section header block)
- SettingsHint.tsx (inline info note)
- SettingsSwitchRow.tsx (label + Radix switch row)
- ObserverStatusBanner.tsx (active/credentials/latest-failure status card)
- Field.tsx (form row container)

Thin wrapper in settings.tsx keeps the ObserverStatusBanner name since the
render loop reads settingsRenderState directly — the extracted component
takes the status as a prop.
