+++ packages/ui/src/tabs/settings/components/SettingsHint.tsx
@@ -1,0 +1,9 @@
+import type { SettingsHintProps } from "../data/types";
+
+export function SettingsHint({ children, hidden = false }: SettingsHintProps) {
+	return (
+		<div className="settings-note" hidden={hidden}>
+			{children}
+		</div>
+	);
+}
