+++ packages/ui/src/tabs/settings/components/SettingsSectionIntro.tsx
@@ -1,0 +1,10 @@
+import type { SettingsSectionIntroProps } from "../data/types";
+
+export function SettingsSectionIntro({ detail, title }: SettingsSectionIntroProps) {
+	return (
+		<div className="settings-section-intro">
+			<div className="settings-section-intro-title">{title}</div>
+			<div className="small settings-section-intro-detail">{detail}</div>
+		</div>
+	);
+}
